# docs/swagger.yaml
definitions:
...
  github_com_KKittyCatik_music_p2p_internal_metadata.TrackMetadata:
    properties:
      artist:
        type: string
      cid:
        type: string
      duration:
        $ref: '#/definitions/time.Duration'
      meta_id:
        description: SHA-256(title + artist + duration)
        type: string
      title:
        type: string
      variants:
        items:
          $ref: '#/definitions/github_com_KKittyCatik_music_p2p_internal_bitrate.Variant'
        type: array
    type: object
  internal_api.ConnectRequest:
    properties:
      multiaddr:
        type: string
    type: object
...
  internal_api.MetadataRequest:
    properties:
      artist:
        type: string
      cid:
        type: string
      duration:
        $ref: '#/definitions/time.Duration'
      title:
        type: string
    type: object
  internal_api.PeerInfo:
    properties:
      id:
        type: string
      score:
        type: number
    type: object
  internal_api.PlayRequest:
    properties:
      cid:
        type: string
    type: object
  internal_api.PlaybackStatus:
    properties:
      buffer_level:
        type: integer
      chunk_index:
        type: integer
      cid:
        type: string
      playing:
        type: boolean
    type: object
...
  internal_api.QueueState:
    properties:
      current:
        $ref: '#/definitions/internal_api.QueueItemResponse'
      upcoming:
        items:
          $ref: '#/definitions/internal_api.QueueItemResponse'
        type: array
    type: object
  internal_api.Response:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
  internal_api.SeekRequest:
    properties:
      chunk_index:
        type: integer
    type: object
...
info:
  contact: {}
paths:
  /dht/provide/{cid}:
    post:
      description: Announce the given CID to the DHT so other peers can find it.
      parameters:
      - description: Track CID
        in: path
        name: cid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: DHT provide
      tags:
      - dht
  /dht/providers/{cid}:
    get:
      description: Find peers that have announced the given CID to the DHT.
      parameters:
      - description: Track CID
        in: path
        name: cid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_api.PeerInfo'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: DHT providers
      tags:
      - dht
...
  /metadata:
    get:
      description: Returns all locally known track metadata entries.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_KKittyCatik_music_p2p_internal_metadata.TrackMetadata'
                  type: array
              type: object
      summary: List metadata
      tags:
      - metadata
    post:
      consumes:
      - application/json
      description: Store and broadcast a new TrackMetadata entry via gossipsub.
      parameters:
      - description: Metadata
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.MetadataRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_KKittyCatik_music_p2p_internal_metadata.TrackMetadata'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Publish metadata
      tags:
      - metadata
  /metadata/{cid}:
    get:
      description: Returns the TrackMetadata for the given CID.
      parameters:
      - description: Track CID
        in: path
        name: cid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_KKittyCatik_music_p2p_internal_metadata.TrackMetadata'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Get metadata by CID
      tags:
      - metadata
  /metadata/search:
    get:
      description: Case-insensitive search over track title and artist fields.
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_KKittyCatik_music_p2p_internal_metadata.TrackMetadata'
                  type: array
              type: object
      summary: Search metadata
      tags:
      - metadata
...
  /peers/{peerID}/score:
    get:
      description: Returns the composite score for the given peer ID.
      parameters:
      - description: Peer ID
        in: path
        name: peerID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_api.PeerInfo'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Peer score
      tags:
      - peers
  /peers/connect:
    post:
      consumes:
      - application/json
      description: Open a connection to the specified peer multiaddr.
      parameters:
      - description: Peer multiaddr
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.ConnectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Connect to peer
      tags:
      - peers
  /playback/play:
    post:
      consumes:
      - application/json
      description: Begin streaming the track identified by the given CID.
      parameters:
      - description: Play request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.PlayRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_api.PlaybackStatus'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Start playback
      tags:
      - playback
  /playback/seek:
    post:
      consumes:
      - application/json
      description: Jump to the given chunk index within the current track.
      parameters:
      - description: Seek request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.SeekRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_api.PlaybackStatus'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Seek playback
      tags:
      - playback
...
  /queue:
    delete:
      description: Remove all upcoming items from the queue (history is preserved).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Clear queue
      tags:
      - queue
    get:
      description: Returns the current item, upcoming items, and play history.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_api.QueueState'
              type: object
      summary: Get queue
      tags:
      - queue
    post:
      consumes:
      - application/json
      description: Add a track to the end of the playback queue.
      parameters:
      - description: Queue item
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.QueueItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_api.QueueState'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Enqueue item
      tags:
      - queue
...
  /queue/insert:
    post:
      consumes:
      - application/json
      description: Insert a track at the given position in the queue (0 = next to
        play).
      parameters:
      - description: Queue item with position
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.QueueItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_api.QueueState'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Insert queue item
      tags:
      - queue
  /status:
    get:
      description: Returns the node's peer ID, listen addresses, connected peer count,
        and uptime.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_api.StatusResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.Response'
      summary: Node status
      tags:
      - node
...
swagger: "2.0"